Fixed naming and a lot of issues
