+++ packages/sdk/src/index.ts
@@ -1,7 +1,7 @@
 // Should be at the top as other modules depend on it
-export * from './spacetimedb.ts';
+export * from './db_connection.ts';
 
 export * from './address.ts';
-export * from './client_db.ts';
+export * from './client_cache.ts';
 export * from './identity.ts';
 export * from './message_types.ts';
